Proxify non-pointer arrays and structs when it makes sense
Remove struct special case in goToLuaFunction.

Add test cases.

Since compound-proxy fields are addressable, they can be proxified, pointer or
not. Besides it is more natural to return a proxy from a proxy field.

The former condition `if isPointerToPrimitive(field)` is a leftover of a former,
more limited GoToLua() function. Same for the condition in struct__newindex.